% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brr.R
\name{as_brr}
\alias{as_brr}
\title{Indicate that a data frame contains balanced repeated replications}
\usage{
as_brr(replications)
}
\arguments{
\item{x}{a data frame or tibble with fits, as produced by \code{brr::brr}}
}
\description{
Alias to \code{brr::new_brr}
}
