\name{brr-package}
\alias{brr-package}
\alias{brr}
\docType{package}
\title{
Bayesian inference on the ratio of two Poisson rates
}
\description{
Implementation of the Bayesian inference for the two independent Poisson samples model, using the semi-conjugate family of prior distributions.
}
\details{
\tabular{ll}{
Package: \tab brr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-10\cr
License: \tab What license is it under?\cr
}
}
\author{
Stéphane Laurent
}
\references{
Laurent, Legrand (2011): A Bayesian framework for the ratio of two Poisson rates in the context of vaccine efficacy trials
}
\keyword{ package }
\keyword{ models }
\seealso{
}
\examples{
}
