% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brr.R
\name{brr}
\alias{brr}
\title{Balanced repeated replications}
\usage{
brr(
  formula,
  statistic,
  data,
  final_weights,
  replicate_weights,
  r = 80,
  .by = "",
  .progress = TRUE,
  ...
)
}
\arguments{
\item{formula}{a formula or a vector of outcomes}

\item{statistic}{statistical function or fitter with signature \code{fitter(data, weights, ...)}}

\item{data}{data}

\item{final_weights}{a tidy selection or a vector of column names}

\item{replicate_weights}{a tidy selection or a vector of column names}

\item{.by}{group observations with \code{.by} when working with long format data where every plausible value has its own row; allows analysis of multiple imputations of the predictors without a performance hit}

\item{.progress}{show a progress bar}

\item{...}{}

\item{replications}{number of replications to perform (for convenience, you can also just pass fewer columns to \code{replicate_weights})}
}
\description{
Balanced repeated replications
}
