% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brr.R
\name{brr_diagnose}
\alias{brr_diagnose}
\title{Diagnose common problems with BRR replications}
\usage{
brr_diagnose(replications)
}
\arguments{
\item{replications}{}
}
\description{
Diagnose common problems with BRR replications
}
