% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brr.R
\name{brr_sd}
\alias{brr_sd}
\title{Imputation, estimation and total standard deviation of balanced repeated replications}
\usage{
brr_sd(replications, perturbation = 0.5, imputation = TRUE, na_rm = FALSE)
}
\arguments{
\item{replications}{BRR replications}

\item{perturbation}{perturbation}

\item{imputation}{Incorporate imputation variance due to plausible values. TRUE by default but can be disabled when there's only a single outcome.}
}
\value{
A list of mean and variance component vectors.
}
\description{
Imputation, estimation and total standard deviation of balanced repeated replications
}
