% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brr.R
\name{brr_var}
\alias{brr_var}
\title{Imputation, estimation and total variance of balanced repeated replications}
\usage{
brr_var(replications, perturbation = 0.5, imputation = TRUE, na_rm = FALSE)
}
\arguments{
\item{perturbation}{perturbation}

\item{imputation}{Incorporate imputation variance due to plausible values. TRUE by default but can be disabled when there's only a single outcome.}

\item{na_rm}{na_rm}

\item{replicates}{BRR replicates}
}
\value{
A list of mean and variance component vectors.
}
\description{
The \code{na_rm} argument should be used sparingly because it may hide model or fit errors,
but it can be particularly useful for repeated cross-sectional analyses of PISA data
from 2000-now, allowing you to specify a model with 10 plausible values even though
2000, 2003, 2006, 2009 and 2012 assessments only include 5.
}
