% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratify.R
\name{clip_strata}
\alias{clip_strata}
\title{Clip strata}
\usage{
clip_strata(strata, upper = 5)
}
\arguments{
\item{strata}{strata with weight adjustment factors, as produced by \code{poststrata}}

\item{upper}{maximum value to which to clip weight adjustment factors}
}
\description{
also known (inaccurately) as trimming and truncating
}
