% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brr.R
\name{coef.brr}
\alias{coef.brr}
\title{Extract model coefficients from the balanced repeated replications of a model fit}
\usage{
\method{coef}{brr}(replications, final_weights = 1, simplify = TRUE)
}
\arguments{
\item{replications}{BRR replications}

\item{final_weights}{the weights index that represents the final weights (usually this is 1)}

\item{simplify}{return a named vector instead of \code{tibble(term, estimate)}}
}
\description{
Extract model coefficients from the balanced repeated replications of a model fit
}
