% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_var.r
\name{con_var}
\alias{con_var}
\title{Data Frame of Numeric Variable
This is an example function named 'con_var'
which extracts numeric column of a data frame and then create a new data frame.}
\usage{
con_var(x)
}
\arguments{
\item{x}{a data frame}
}
\value{
a dataframe of numeric column
}
\description{
Data Frame of Numeric Variable
This is an example function named 'con_var'
which extracts numeric column of a data frame and then create a new data frame.
}

