% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brr.R
\name{confint.brr}
\alias{confint.brr}
\title{Credibility Intervals}
\usage{
\method{confint}{brr}(model, conf = 0.95, intervals = "all", ...)
}
\arguments{
\item{model}{a \code{link[Brr]{brr}} object}

\item{conf}{confidence level}

\item{intervals}{a character vector, the intervals to be returned}

\item{...}{other aguments passed to \code{\link{credibility_intervals}} (currently nothing)}
}
\value{
A list of confidence intervals
}
\description{
Wrapper to \code{\link{credibility_intervals}}
}
\examples{
model <- Brr(x=10, y=10, S=100, T=100)
confint(model)
}

