% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attribute_Value_Frequency.r
\name{df.AVF}
\alias{df.AVF}
\title{The Attribute Value Frequency of Categorical variable}
\usage{
df.AVF(x, name = "AVF")
}
\arguments{
\item{x}{a data frame to be which have both numerical and categorical variable}

\item{name}{it the name of the new data frame to be given which will include he AVF score for each row}
}
\value{
it returns a data frame with avf score for outliers detection
}
\description{
This is an example function name 'df.AVF'
The function will calculate the frequency of each unique variable in every column over the row
and sum the total number of count for each unique value in row which then will be divided by the
total number of categorical variable
}
