% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor.r
\name{df.factor}
\alias{df.factor}
\title{Factor of Character Column
This is an example function named 'df.factor'
which convert a character column to factor.}
\usage{
df.factor(x)
}
\arguments{
\item{x}{the dataframe}
}
\value{
the dataframe character column now converted into factor
}
\description{
Factor of Character Column
This is an example function named 'df.factor'
which convert a character column to factor.
}
