% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.r
\name{df.ks_stats}
\alias{df.ks_stats}
\title{Kolmogorov-Smirnov test FKA ks-test function}
\usage{
df.ks_stats(True_Label = NULL, Prob = NULL, Plot = TRUE)
}
\arguments{
\item{True_Label}{this the actual target variable of the data}

\item{Prob}{this the predicted probability of the target variable using some model}

\item{Plot}{this is a logical value which tell to plot the ks graph in the console}
}
\value{
this function will return a list of values like the ks table, decile and the ks ststistics
}
\description{
This function will calculate the KS value of the given propability against the true label and also provide the decile under which it falls
}
