% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Binning.r
\name{DF.Replace.Bin}
\alias{DF.Replace.Bin}
\title{Create Binning variable from  Information Table}
\usage{
DF.Replace.Bin(X, y, Dependent = NULL)
}
\arguments{
\item{X}{it's a data frame that is used to create the Information table from Information package.}

\item{y}{it's a information table that is created by information package.}

\item{Dependent}{it's a dependent variable that is used to classify good or bad.}
}
\value{
it will return a data frame with binning variables from information table.
}
\description{
The Function DF.Replace.Bin will create a data frame of binning variables as seen in Information table,
it takes the argument dataframe, information table and dependent variable and then create seperate df with all
the binning variables.
}
