% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOE.r
\name{DF.Replace.WOE}
\alias{DF.Replace.WOE}
\title{Data Frame with Weight Of Evidence values from Information package}
\usage{
DF.Replace.WOE(X, y, Dependent = NULL)
}
\arguments{
\item{X}{is a data sets which is used to create the Information table using information package.}

\item{y}{is the Information table which is created using information package.}

\item{Dependent}{is the dependent binary variable which is used to classify good and bad.}
}
\value{
It will return the data frame with WOE values
}
\description{
This is an function name 'DF.Replace.WOE', which will extract the values of WOE from Information
list which can be generated using information package and replace the original values with the
WOE values.
}
