% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorecard.r
\name{df.score}
\alias{df.score}
\title{Application ScoreCard for Credit Risk Model}
\usage{
df.score(Applicant = NULL, PD = NULL, True_Label = NULL, score = NULL,
  pdo = NULL, Odd = NULL)
}
\arguments{
\item{Applicant}{this the unique applicant id for which we are require to create a score card}

\item{PD}{this is the predicted propability which is generated by using predict() function on logistic model.}

\item{True_Label}{this will the actual label for which prediction is generated either from train or test data sets.}

\item{score}{this is the maximum score that is given to the applicant who is very active in paying the debts owe to him.}

\item{pdo}{this is the point to double the odds}

\item{Odd}{this the good to bad odd}
}
\value{
the function will return a data frame with the scores along with all the above parameters
}
\description{
This Function will create a application scorecard from the predicted probability of the logistic regression model.
Suppose we have to Build an application scorecard with the good to bad odds of 10 to 1 at a score of 400 doubling every 20 points.
So here Odd will be 10, pdo will be 20 and score will be 400.
}
