% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brr.R
\name{-.brr}
\alias{-.brr}
\title{Subtract sets of replications from each other}
\usage{
\method{-}{brr}(e1, e2)
}
\arguments{
\item{e1}{lefthand set of replications}

\item{e2}{righthand set of replications}
}
\description{
Useful to compare scores between countries or between cycles when passed on to
\code{confint}
}
