% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratify.R
\name{drop_na_strata}
\alias{drop_na_strata}
\title{Drop strata that include NA factor levels}
\usage{
drop_na_strata(strata)
}
\arguments{
\item{strata}{strata with weight adjustment factors, as produced by \code{poststrata}}
}
\description{
Drop strata that include NA factor levels
}
