% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadamard.R
\name{fay_tibble}
\alias{fay_tibble}
\title{Balanced repeated replication weight matrix with Fay's correction}
\usage{
fay_tibble(
  ncol = 80,
  nrow = ncol,
  scale = 0.5,
  location = 1,
  weights = 1,
  prefix = "w_student_r"
)
}
\arguments{
\item{ncol}{width of the matrix}

\item{nrow}{length of the matrix}

\item{scale}{scale by which to multiply the original entries (-1 and 1)}

\item{location}{location by which to shift the scaled entries}

\item{weights}{sampling weights or a single weight to further adjust the weights of each row}

\item{prefix}{}
}
\description{
This replication weight matrix generalizes the Hadamard matrix with arbitrary
weights (or weight adjustment factors) instead of -1 and 1. The default
scale and location parameters produce weights of 0.5 and 1.5, as with Fay's
method.

Using the \code{weights} argument,these replications weights can then further be
multiplied with the sampling weights so they can be used in a balanced
repeated replication.

If \code{nrow} is different from \code{ncol}, or if sampling weights are provided,
the matrix will be repeated or abbreviated to attain the desired amount of
rows.
}
