% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{GammaInverseBetaDist}
\alias{GammaInverseBetaDist}
\alias{dGIB}
\alias{rGIB}
\alias{summary_GIB}
\title{Gamma-Inverse Beta distribution}
\usage{
dGIB(x, a, alpha, beta, rho)

rGIB(n, a, alpha, beta, rho)

summary_GIB(a, alpha, beta, rho, type = "list", ...)
}
\arguments{
\item{x}{vector of quantiles}

\item{a}{non-negative shape parameter of the Gamma distribution}

\item{alpha,beta}{non-negative shape parameters of the mixing Beta distribution}

\item{n}{number of observations to be simulated}
}
\value{
\code{dGIB} gives the density, \code{rGIB} samples from the distribution,
and \code{summary_GIB} returns a summary of the distribution.
}
\description{
Density and random  generation for the Gamma-Inverse Beta distribution
with shape parameters \code{a}, \code{alpha}, \code{beta} and scale parameter \code{rho}.






}
\details{
This is the mixture distribution obtained by sampling a value \eqn{b} from a Beta distribution
with shape parameters \eqn{\beta}, \eqn{\alpha}
and then sampling a Gamma distribution with shape \eqn{a} and rate \eqn{\rho/b}.
}
\note{
\code{GammaInverseBetaDist} is a generic name for the functions documented.
}
\examples{
curve(dGIB(x,3,4,2,2.5), from=0, to=3)
sims <- rgamma(100000, 3, 2.5/rbeta(100000,2,4))
lines(density(sims, from=0), col="red")
lines(density(rGIB(100000, 3, 4, 2, 2.5), from=0), col="green")
mean(sims); var(sims)
summary_GIB(3,4,2,2.5,type="pandoc")
}

