% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{ginv}
\alias{ginv}
\title{Generalized (Pseudo-)Inverse}
\usage{
ginv(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{matrix}

\item{tol}{tolerance}
}
\description{
\code{ginv} is inspired by code in William Revelle's \verb{psych package}
}
