% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadamard.R
\name{hadamard_matrix}
\alias{hadamard_matrix}
\title{Hadamard matrix}
\usage{
hadamard_matrix(order)
}
\arguments{
\item{order}{order of the matrix}
}
\description{
Compute a Hadamard matrix with an order of any power of two, plus
a handful of other useful sizes including 20, 60, 80, 140, 180, 200.
}
