% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkt_seg.r
\name{imp_seg}
\alias{imp_seg}
\title{Segmentation of Market based on Profit}
\usage{
imp_seg(x = "Dataframe", c = "aggregate_column", d = "Date_col",
  m = "no of month to forecast",
  v = "Profit column for covariance calculation",
  n = "Top n profitable segments")
}
\arguments{
\item{x}{it is the data frame to calculate the market segments}

\item{c}{it is the column such as sales, quantity, profit etc which is required to aggregate based on date}

\item{d}{it is the date column for time series}

\item{m}{it is the no of months that needs to be seperated for validation purpose}

\item{v}{it is the profit column or can be any column based on which covariance is to be calculated}

\item{n}{it is the top n segments that is require for predicting sales and quantity for that segments}
}
\value{
this will return a list of top segments, the train and test time series
}
\description{
This function is created to calculate the top market segments base on minumum covariance of the profit made on each segments
}
