% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brr.R
\name{inference.brr}
\alias{coef.brr}
\alias{confint.brr}
\alias{inference.brr}
\alias{print.coef.brr}
\alias{print.confint.brr}
\title{Credibility intervals and estimates}
\usage{
\method{confint}{brr}(model, conf = 0.95, intervals = "all", ...)

\method{print}{confint.brr}(x, style = "grid")

\method{coef}{brr}(model, parameter = "phi", ...)

\method{print}{coef.brr}(x)
}
\arguments{
\item{model}{a \code{\link[Brr]{brr}} object}

\item{conf}{confidence level}

\item{intervals}{a character vector, the intervals to be returned}

\item{...}{other aguments passed to \code{\link{brr_intervals}} or \code{\link{brr_estimates}}}

\item{style}{the style of the table to print
(passed to  \code{\link[=pander]{pandoc.table.return}})}
}
\value{
\code{confint.brr} returns a list of confidence intervals
}
\description{
Get credibility intervals and estimates from a \code{brr} object






}
\details{
\code{confint.brr} is a wrapper to \code{\link{brr_intervals}} and
\code{coef.brr} is a wrapper to \code{\link{brr_estimates}}
}
\examples{
model <- Brr(x=10, y=10, S=100, T=100)
confint(model)
coef(model)
}

