% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intrinsic2.R
\name{intrinsic2_discrepancy}
\alias{intrinsic2_discrepancy}
\title{Second intrinsic discrepancy}
\usage{
intrinsic2_discrepancy(phi0, phi, a, b, S, T)
}
\arguments{
\item{phi0}{the proxy value of \code{phi}}

\item{phi}{the true value of the parameter}

\item{a,b,}{the parameters of the prior Gamma distribution on \eqn{\mu}}

\item{S,T}{sample sizes}
}
\value{
A number, the intrinsic discrepancy from \code{phi0} to \code{phi}.
}
\description{
Intrinsic discrepancy from \code{phi0} to \code{phi} in the marginal model.
}

