% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intrinsic.R
\name{intrinsic_bounds}
\alias{intrinsic_bounds}
\title{Intrinsic interval}
\usage{
intrinsic_bounds(x, y, S, T, a, b, c, d, conf, parameter = "phi",
  subd = 10000, tol = 1e-08)
}
\description{
Intrinsic interval
}
\examples{
a<-0.5; b<-0; c<-1/2; d<-0; S<-100; T<-S; x<-0; y<-20
intrinsic_estimate(x, y, S, T, a, b, c, d)
bounds <- intrinsic_bounds(x, y, S, T, a, b, c, d, conf=0.95); bounds
ppost_phi(bounds[2], a, b, c, d, S, T,  x, y)- ppost_phi(bounds[1], a, b, c, d, S, T, x, y)
}

