% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intrinsic.R
\name{intrinsic_discrepancy}
\alias{intrinsic_discrepancy}
\title{Intrinsic discrepancy}
\usage{
intrinsic_discrepancy(phi0, mu, phi, S, T)
}
\arguments{
\item{phi0}{the proxy value of \code{phi}}

\item{mu,phi}{the true values of the parameters}

\item{S,T}{sample sizes}
}
\value{
A number, the intrinsic discrepancy from \code{phi0} to \code{(mu,phi)}.
}
\description{
Intrinsic discrepancy from \code{phi0} to \code{(mu,phi)}.
}

