% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intrinsic.R
\name{intrinsic_estimate}
\alias{intrinsic_estimate}
\title{Intrinsic estimate}
\usage{
intrinsic_estimate(x, y, S, T, a, b, c, d, subd = 10000, tol = 1e-08)
}
\description{
Intrinsic estimate
}
\examples{
a<-0.5; b<-0; c<-1/2; d<-0; S<-100; T<-S; x<-0; y<-20
intrinsic_estimate(x, y, S, T, a, b, c, d)
}

