% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intrinsic.R
\name{intrinsic_H0}
\alias{intrinsic_H0}
\title{Unilateral intrinsic cost}
\usage{
intrinsic_H0(phi.star, alternative, x, y, S, T, a = 0.5, b = 0, c = 0.5,
  d = 0, subd = 10000, tol = 1e-06)
}
\description{
alternative="less" for  H1: phi0 < phi.star;
alternative="gretaer" for  H1: phi0 > phi.star
}

