% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intrinsic.R
\name{intrinsic_phi0}
\alias{intrinsic_phi0}
\title{Posterior intrinsic cost}
\usage{
intrinsic_phi0(phi0, x, y, S, T, a = 0.5, b = 0, c = 0.5, d = 0,
  subd = 10000, tol = 1e-06)
}
\description{
Posterior intrinsic cost
}
\examples{
a<-0.5; b<-0; c<-1/2; d<-0; S<-100; T<-S; x<-0; y<-20
intrinsic_phi0(0.5, x, y, S, T, a, b, c, d)
}

