% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{invisibly}
\alias{invisibly}
\title{Make any function behave like \code{identity}, with side effects}
\usage{
invisibly(f)
}
\arguments{
\item{f}{a function without parameters, run for its side effects}
}
\value{
a wrapped function that returns it argument, like \code{identity}
}
\description{
Make any function behave like \code{identity}, with side effects
}
\examples{
invisibly(\() print("hello world"))(4) ^ 2
}
