% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brr.R
\name{is_brr}
\alias{is_brr}
\title{Test whether an object is a collection of replications}
\usage{
is_brr(x)
}
\arguments{
\item{x}{any object}
}
\description{
Test whether an object is a collection of replications
}
