% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{label_vector}
\alias{label_vector}
\title{Label a vector}
\usage{
label_vector(v, label)
}
\arguments{
\item{v}{vector}

\item{label}{name}
}
\value{
a vector with a label attribute
}
\description{
In conjunction with \code{brr::redirect_weights}, allows the use of \code{brr:brr}
fitters or other functions that have a different name for their weights
argument, and/or fitters that take a weights column name rather than the
actual weights.
}
