% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{marginal_to_partial}
\alias{marginal_to_partial}
\title{Obtain partial correlations from a correlation matrix}
\usage{
marginal_to_partial(x)
}
\arguments{
\item{x}{matrix}
}
\description{
\code{marginal_to_partial} is inspired by code in William Revelle's \verb{psych package}
}
