% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brr.R, R/brrl.R
\name{new_brr}
\alias{new_brr}
\title{Indicate that a data frame contains balanced repeated replications}
\usage{
new_brr(fits)

new_brr(fits)
}
\arguments{
\item{x}{a data frame or tibble with fits, as produced by \code{brr::brr}}
}
\description{
Indicate that a data frame contains balanced repeated replications

Indicate that a data frame contains balanced repeated replications
}
