% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadamard.R
\name{paley_matrix}
\alias{paley_matrix}
\title{Hadamard matrix with Paley's method}
\usage{
paley_matrix(order)
}
\arguments{
\item{order}{order of the matrix}
}
\description{
Paley's method works for orders that are one larger than a prime number,
where that prime divided by 4 has a remainder of 3, so e.g. 80 because
79 is a prime number and 79 divided by 4 has a remainder of 3.

Adapted from code originally written Appavoo Dhandapani and Revan Siddesha.
}
