% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brr.R
\name{plot.brr}
\alias{plot.brr}
\title{plot brr}
\usage{
\method{plot}{brr}(brr, what = "summary")
}
\description{
plot brr
}
\examples{
model <- Brr(a=2, b=3)
plot(model)
plot(model, dprior(mu))
plot(model, pprior(mu))
model <- model(c=4, d=6, S=10, T=10)
plot(model)
plot(model, dprior(phi))
plot(model, dprior(x))
}

