% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratify.R
\name{poststrata}
\alias{poststrata}
\title{Calculate weight adjustment factors so that the proportion of each strata in \code{df} matches the proportion in \code{df_target}}
\usage{
poststrata(df, df_target, factors, weights = NULL)
}
\arguments{
\item{df}{data frame for which to calculate the adjustment factors}

\item{df_target}{data frame that serves as the target or baseline}

\item{factors}{factor variables on which to poststratify}

\item{weights}{column name of a column that contains observation weights}
}
\description{
Calculate weight adjustment factors so that the proportion of each strata in \code{df} matches the proportion in \code{df_target}
}
