% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poststratify.R
\name{poststratify}
\alias{poststratify}
\title{Poststratify a dataset with strata weight adjustment factors}
\usage{
poststratify(df, strata, name = "weight")
}
\arguments{
\item{df}{a data frame}

\item{strata}{strata with weight adjustment factors, as produced by \code{poststrata(df, ...)}}

\item{name}{rename the weight adjustment factors to \code{name}}
}
\description{
Poststratify a dataset with strata weight adjustment factors
}
