% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{pull_weighted_mean}
\alias{pull_weighted_mean}
\title{Weighted mean with a formula interface}
\usage{
pull_weighted_mean(formula, data, ...)
}
\arguments{
\item{formula}{formula of the form \code{outcome ~ 1}, with a single outcome and no left-hand side variables}

\item{data}{data}

\item{...}{additional arguments, passed onto \code{weighted_mean}, such as \code{na_rm}}

\item{weights}{a vector of weights}
}
\description{
See \code{brr::formulaic} for more details. Useful because \code{brr::brr} expects a statistic
with a formula interface.
}
