% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{reliability_to_variance}
\alias{reliability_to_variance}
\title{Convert from reliability (complement of the error proportion of the variance) to the error variance}
\usage{
reliability_to_variance(x, r, ...)
}
\arguments{
\item{x}{x}

\item{r}{r}

\item{...}{}
}
\description{
In psychometrics, reliability is commonly expressed as 1 minus the variance of the error to the total variance of the variable,
but software packages typically expect an error variance instead.
}
