% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brr.R
\name{summary.brr}
\alias{summary.brr}
\title{Summarize balanced repeated replications}
\usage{
\method{summary}{brr}(replications)
}
\arguments{
\item{replications}{}
}
\description{
This function is a placeholder and has not yet been implemented.
In the meanwhile, use \code{coef} and \code{confint}.
}
