% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{summary_nbinom}
\alias{summary_nbinom}
\title{Summary of a Negative Binomial distribution}
\usage{
summary_nbinom(a, p, type = "list", ...)
}
\arguments{
\item{a,p}{parameters of the negative binomial distribution}

\item{type}{type of the output: \code{"list"} to return a list, \code{"pandoc"} to print a table}

\item{...}{arguments passed to \code{\link[=pander]{pander.data.frame}}}
}
\description{
Mode, mean, variance, and quartiles for a Negative Binomial distribution
with shape parameter \code{a} and probability parameter \code{p}.
}
\examples{
summary_nbinom(a=2, p=0.4, type="pandoc", style="rmarkdown")
}

