% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hadamard.R
\name{sylvester_matrix}
\alias{sylvester_matrix}
\title{Hadamard matrix of order 2^k with Sylvester's method}
\usage{
sylvester_matrix(order)
}
\arguments{
\item{order}{order of the matrix}
}
\description{
Hadamard matrix of order 2^k with Sylvester's method
}
