% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{tidy.aggregate}
\alias{tidy.aggregate}
\title{Tidy an aggregate object}
\usage{
\method{tidy}{aggregate}(results, index = 1)
}
\arguments{
\item{results}{a tibble with aggregated outcomes as produced by \code{weighted_aggregate}}

\item{index}{the estimate to extract, if there is more than one outcome}
}
\description{
Tidy an aggregate object
}
