% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brr.R
\name{tidy.brr}
\alias{tidy.brr}
\title{Tidy a BRR object (experimental)}
\usage{
\method{tidy}{brr}(replications, na_rm = FALSE)
}
\arguments{
\item{replications}{A brr object, as produced by \code{brr::brr}}
}
\description{
Tidy a BRR object (experimental)
}
