% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{unsafely}
\alias{unsafely}
\title{Wrap a function to match the return value of \code{purrr:safely}, without capturing errors}
\usage{
unsafely(f)
}
\arguments{
\item{f}{a function to wrap}
}
\description{
Wrap a function to match the return value of \code{purrr:safely}, without capturing errors
}
