% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brr.R
\name{VarCorr.brr}
\alias{VarCorr.brr}
\title{Imputation, estimation and total variance of balanced repeated replications}
\usage{
VarCorr.brr(replications, perturbation = 0.5, final_weights = 1)
}
\arguments{
\item{replications}{BRR replications}

\item{perturbation}{perturbation}

\item{final_weights}{the weights index that represents the final weights (usually this is 1)}
}
\value{
a list of mean and variance component vectors
}
\description{
Imputation, estimation and total variance of balanced repeated replications
}
