% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted_aggregate}
\alias{weighted_aggregate}
\title{Compute weighted summary statistics of data subsets (or the entire dataset) for one or more outcome variables}
\usage{
weighted_aggregate(formula, statistic, data, weights, ...)
}
\arguments{
\item{formula}{formula}

\item{statistic}{statistical function or fitter with signature \code{fitter(data, weights, ...)}}

\item{data}{data}

\item{weights}{a vector of weights}

\item{...}{additional arguments, passed onto `statistic``}
}
\description{
This function is powerful but can get a bit slow. To compute an aggregate for a single outcome
without any groups to condition on, consider using \code{brr::formulaic} instead, e.g.
\code{formulaic(weighted_mean)(outcome ~ 1, data, weights)} to create a formula interface for
a simple statistical function.
}
\examples{
df <- tibble(x=1:10, y=11:20, w=1 + 1:10/10, u=rep(1, 10), g=rep(c(1,2), each=5))
weighted_aggregate(x + y ~ g, weighted_mean, df, df$w)
# or more succinct...
weighted_mean_by <- partial(weighted_aggregate, statistic=weighted_mean)
weighted_mean_by(x + y ~ g, df, df$w)
}
