% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted_cor_matrix}
\alias{weighted_cor_matrix}
\title{Weighted correlation matrix}
\usage{
weighted_cor_matrix(data, weights, center = TRUE, method = "unbiased")
}
\arguments{
\item{data}{data}

\item{weights}{a non-negative and non-zero vector of weights for each observation. Its length must equal the number of rows of x}

\item{center}{either a logical or a numeric vector specifying the centers to be used when computing correlations}

\item{method}{string specifying how the result is scaled, \code{unbiased} or \code{ML}}
}
\description{
Weighted correlation matrix
}
