% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted_cov}
\alias{weighted_cov}
\title{Weighted covariance}
\usage{
weighted_cov(x, y, weights, na_rm = FALSE, scale = FALSE)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame}

\item{y}{a numeric vector, matrix or data frame}

\item{weights}{a vector of weights for each observation}

\item{na_rm}{remove observations for which either the weight or one or both variables are NA}

\item{scale}{calculate a correlation instead}
}
\description{
Weighted covariance
}
