% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted_cov_matrix}
\alias{weighted_cov_matrix}
\title{Weighted covariance matrix}
\usage{
weighted_cov_matrix(
  data,
  weights,
  center = TRUE,
  method = "unbiased",
  na_rm = FALSE
)
}
\arguments{
\item{data}{data}

\item{weights}{a non-negative and non-zero vector of weights for each observation. Its length must equal the number of rows of x}

\item{center}{either a logical or a numeric vector specifying the centers to be used when computing covariances}

\item{method}{string specifying how the result is scaled, \code{unbiased} or \code{ML}}
}
\description{
Weighted covariance matrix
}
