% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted_mean}
\alias{weighted_mean}
\title{Weighted mean}
\usage{
weighted_mean(x, weights, na_rm = FALSE, ...)
}
\arguments{
\item{x}{a data vector}

\item{weights}{a weights vector}

\item{na_rm}{na_rm}

\item{...}{}
}
\description{
Weighted mean
}
