% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted_mean_by}
\alias{weighted_mean_by}
\title{Weighted mean conditional on zero or more groups}
\usage{
weighted_mean_by(...)
}
\arguments{
\item{...}{additional arguments, passed onto \code{weighted_mean}, such as \code{na_rm}}

\item{formula}{formula}

\item{data}{data}

\item{weights}{a vector of weights}
}
\description{
Weighted mean conditional on zero or more groups
}
