% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted_median_by}
\alias{weighted_median_by}
\title{Weighted median conditional on zero or more groups}
\usage{
weighted_median_by(...)
}
\arguments{
\item{...}{additional arguments, passed onto \code{weighted_median}, such as \code{na_rm}}

\item{formula}{formula}

\item{data}{data}

\item{weights}{a vector of weights}
}
\description{
Weighted median conditional on zero or more groups
}
