% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted_partial_cor_matrix}
\alias{weighted_partial_cor_matrix}
\title{Weighted partial correlation matrix}
\usage{
weighted_partial_cor_matrix(
  data,
  weights,
  center = TRUE,
  method = "unbiased",
  na_rm = FALSE
)
}
\arguments{
\item{data}{matrix}

\item{weights}{weights}

\item{center}{center}

\item{method}{method}

\item{na_rm}{na_rm, equivalent to \code{cor(use = "complete.obs")}, as pairwise complete etc. does not make sense for partial correlations}
}
\description{
Weighted partial correlation matrix
}
