% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted_quantile_hd82}
\alias{weighted_quantile_hd82}
\title{Distribution-free weighted quantile estimator}
\usage{
weighted_quantile_hd82(x, weights, probs, na_rm = FALSE)
}
\arguments{
\item{x}{a data vector}

\item{weights}{a weights vector}

\item{probs}{requested quantiles}

\item{na_rm}{na_rm}
}
\description{
Authored by Andrey Akinshin according to the proposal in Harrell & Davis 1982
}
