% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted_r2}
\alias{weighted_r2}
\title{Weighted variance explained}
\usage{
weighted_r2(formula, data, weights, ...)
}
\arguments{
\item{formula}{formula}

\item{data}{data}

\item{weights}{weights}

\item{...}{}
}
\description{
Weighted variance explained
}
