% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted_reliability_to_variance}
\alias{weighted_reliability_to_variance}
\title{Convert from weighted reliability (complement of the error proportion of the variance) to the weighted error variance}
\usage{
weighted_reliability_to_variance(x, weights, r, ...)
}
\arguments{
\item{x}{a data vector}

\item{weights}{a weights vector}

\item{r}{reliability between 0.0 and 1.0}

\item{...}{}
}
\description{
In psychometrics, reliability is commonly expressed as 1 minus the variance of the error to the total variance of the variable,
but software packages typically expect an error variance instead.
}
