% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted_scale}
\alias{weighted_scale}
\title{Weighted scaling and centering of a vector}
\usage{
weighted_scale(x, weights, na_rm = TRUE)
}
\arguments{
\item{x}{a data vector}

\item{weights}{a weights vector}

\item{na_rm}{in line with \code{base::scale}, removes NAs by default}
}
\description{
Weighted scaling and centering of a vector
}
