% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted_scale_by}
\alias{weighted_scale_by}
\title{Scaling and centering of a vector by the weighted mean and sd of a target vector}
\usage{
weighted_scale_by(x, x_target, weights, na_rm = TRUE)
}
\arguments{
\item{x}{a data vector}

\item{x_target}{a target data vector}

\item{weights}{a target weights vector}

\item{na_rm}{in line with \code{base::scale}, removes NAs by default}
}
\description{
Sometimes this is useful to make sure that a bunch of different vectors all receive the same rescaling
}
