% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted_sd}
\alias{weighted_sd}
\title{Weighted standard deviation}
\usage{
weighted_sd(x, weights, na_rm = FALSE)
}
\arguments{
\item{x}{a data vector}

\item{weights}{a weights vector}

\item{na_rm}{remove NA values}
}
\description{
Weighted standard deviation
}
