% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted_sum}
\alias{weighted_sum}
\title{A weighted sum}
\usage{
weighted_sum(x, weights, na_rm = FALSE)
}
\arguments{
\item{weights}{a weights vector}

\item{na_rm}{na_rm}

\item{data}{a data vector}
}
\description{
Really just a sum of products. Weights are not normalized but used as-is.
}
