% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted_var}
\alias{weighted_var}
\title{Weighted variance}
\usage{
weighted_var(x, weights, na_rm = FALSE)
}
\arguments{
\item{x}{a data vector}

\item{weights}{a weights vector}

\item{na_rm}{remove NA values}
}
\description{
Authored by Gavin Simpson.
}
