\name{brr-package}
\alias{brr-package}
\alias{brr}
\docType{package}
\title{
Bayesian inference on the ratio of two Poisson rates
}
\description{
Implementation of the Bayesian inference for the two independent Poisson samples model, using the semi-conjugate family of prior distributions. 

To learn more, start with the vignettes:
\code{browseVignettes(package = "brr")}
}
\details{
\tabular{ll}{
Package: \tab brr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-10\cr
License: \tab GPL-2\cr
}
}

\author{
Stéphane Laurent
}

\references{
S. Laurent, C. Legrand (2011): 
A Bayesian framework for the ratio of two Poisson rates in the context of vaccine efficacy trials.
ESAIM, Probability \& Statistics (2011). Published online by Cambridge University Press
(doi:10.1051/ps/2010018)

S. Laurent (2012):  
Some Poisson mixtures  distributions with a hyperscale parameter. 
Brazilian Journal of Probability and Statistics 26 (2012), 265--278.

S. Laurent (2012):
Intrinsic Bayesian inference on a Poisson rate and on the ratio of two Poisson rates. 
Journal of Statistical Planning and Inference 142 (2012), 2656--2671.
}
\keyword{ package }
\keyword{ models }
\seealso{
}
\examples{
}
