% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brr.R
\name{brr}
\alias{brr}
\title{Balanced repeated replications}
\usage{
brr(
  formula,
  statistic,
  data,
  final_weights,
  replicate_weights,
  r = 80,
  .progress = TRUE,
  .select = "tidy",
  ...
)
}
\arguments{
\item{formula}{formula}

\item{statistic}{statistical function or fitter with signature \code{fitter(data, weights, ...)}}

\item{data}{data}

\item{final_weights}{depending on \code{.select} either a tidy selection (the default);
a vector of column names; or a vector, one-column data frame or one-column tibble}

\item{replicate_weights}{depending on \code{.select} either a tidy selection (the default);
a vector of column names; or a data frame or tibble}

\item{.progress}{show a progress bar}

\item{.select}{select weights from \code{data} using selection helpers such as \code{starts_with}, \code{matches} etc. from the tidyselect package (tidy); using column names (names); or by passing the data itself (noop)}

\item{...}{}

\item{replicates}{number of replications to perform (for convenience, you can also just pass fewer columns to \code{replicate_weights})}
}
\description{
Balanced repeated replications
}
