% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brr.R
\name{Brr}
\alias{Brr}
\title{Set up}
\usage{
Brr(...)
}
\description{
Set up
}
\examples{
model <- Brr(a=2, b=3)
model()
# add parameters
model <- model(c=4, d=5)
model()
# replace parameters
model <- model(a=10, b=11)
model()
}

