% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brrl.R
\name{brrl}
\alias{brrl}
\title{Balanced repeated replications}
\usage{
brrl(
  statistic,
  data,
  weights,
  conditions,
  i = NULL,
  r = NULL,
  .progress = TRUE,
  .verbose = TRUE
)
}
\arguments{
\item{statistic}{statistical function or fitter with signature \code{fitter(data, weights)}}

\item{data}{plausible values and covariates in long format, one imputation per row}

\item{weights}{weights shared by all imputations, of which the first weight will be used as final weight}

\item{i}{number of imputations to to analyze, if only a subset of \code{data} should be used}

\item{r}{number of replications to perform, if only a subset of \code{weights} should be used}

\item{.progress}{show a progress bar}

\item{...}{}
}
\description{
\code{brrl} allows multiple imputation of covariates (not just plausible values)
without performance overhead relative to the analysis of a single imputation
as with the original \code{brr} function.

Because \code{weights} are shared by all imputations, they are typically provided
by a data frame with 1/5th or 1/10th the amount of rows as the \code{data} set.

See \code{pisa.rx.parquet} for a multiply imputed dataset that is compatible with
\code{brrl}, or alternatively use \code{pivot_longer(..., names_pattern = 'pv(\\\\d+)(math)', names_to = c('i', '.value'))}
on a wide format PISA dataset and merge these plausible values with imputations
such as those generated with \code{mice}.
}
