% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/con_var.r
\name{con_var}
\alias{con_var}
\title{Data Frame of Numeric Variable
This is an example function named 'con_var'
which extracts numeric column of a data frame and then create a new data frame.
The data frame can be given any name using name attribute, by default the name is "Numeric_VAR".}
\usage{
con_var(x, name = "Numeric_VAR")
}
\arguments{
\item{x}{a data frame}

\item{name}{the name of the new numeric data frame}
}
\value{
a dataframe of numeric column
}
\description{
Data Frame of Numeric Variable
This is an example function named 'con_var'
which extracts numeric column of a data frame and then create a new data frame.
The data frame can be given any name using name attribute, by default the name is "Numeric_VAR".
}
