\name{df.factor}
\alias{df.factor}

\title{
Convert Character to Factor
}
\description{
The function df.factor takes the argument of a data frame and find character attributes, the matched attributes is then converted to class factor, the output of the function is a data frame and can be store in a new data frame.
}
\usage{
df.factor(x)
}

\arguments{
  \item{x}{
A Data Frame
}
}

\value{
Convert Character attribute of to Factor
}

\author{
Prabhakar Srivastav
}





\examples{
## check the structure of dataframe
## df.factor(dataframe)
}


