% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/intrinsic.R
\name{intrinsic_bounds}
\alias{intrinsic_bounds}
\title{Intrinsic credible interval}
\usage{
intrinsic_bounds(x, y, S, T, a, b, c, d, conf = 0.95, parameter = "phi",
  subd = 1000, tol = 1e-08)
}
\arguments{
\item{S,T}{sample sizes}

\item{a,b,c,d}{Prior parameters}

\item{conf}{credibility level}

\item{parameter}{relative risk \code{"phi"} or vaccine efficacy \code{"VE"}}

\item{subd}{number of subdividisions passed to the \code{\link{integrate}} function
through the \code{\link{intrinsic_phi0}} function}
}
\value{
a vector with two numbers, the bounds of the intrinsic credible interval
}
\description{
Intrinsic credible interval about the relative risk or the vaccine efficacy
}
\examples{
a<-0.5; b<-0; c<-1/2; d<-0; S<-100; T<-S; x<-0; y<-20
intrinsic_estimate(x, y, S, T, a, b, c, d)
bounds <- intrinsic_bounds(x, y, S, T, a, b, c, d, conf=0.95); bounds
ppost_phi(bounds[2], a, b, c, d, S, T,  x, y)- ppost_phi(bounds[1], a, b, c, d, S, T, x, y)
}

