% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brr.R
\name{plot.brr}
\alias{plot.brr}
\title{plot brr}
\usage{
\method{plot}{brr}(model, what = "summary")
}
\arguments{
\item{model}{an object of class \code{brr} (see \code{\link{Brr}})}

\item{what}{\code{"summary"} to plot automatically the priors on \code{mu} and \code{phi}
and the posterior on \code{phi}, or an expression like \code{dprior(mu)} for a specific plot (see examples)}
}
\description{
plot brr
}
\examples{
model <- Brr(a=2, b=3)
plot(model)
plot(model, dprior(mu))
plot(model, pprior(mu))
model <- model(c=4, d=6, S=10, T=10)
plot(model)
plot(model, dprior(phi))
plot(model, dprior(x))
}

