% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brr.R
\name{summary.brr}
\alias{summary.brr}
\title{Summary brr}
\usage{
\method{summary}{brr}(brr, phi0 = 1, hypothesis = "greater",
  table.style = "grid")
}
\description{
Summary brr
}
\examples{
model <- Brr()
summary(model)
model <- Brr(x=3, y=4)
summary(model)
model <- Brr(a=2, b=4, T=10)
summary(model)
model <- model(a=2, b=4, c=3, d=5, S=10, T=10)
summary(model)
model <- model(x=5, y=10)
summary(model)
}

