% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{summary_gamma}
\alias{summary_gamma}
\title{Summary of a Gamma distribution}
\usage{
summary_gamma(a, b, type = "list", ...)
}
\arguments{
\item{a,b}{Shape and rate parameters.}
}
\description{
Mode, mean, variance, and quartiles for a Gamma distribution
with shape parameter \code{a} and rate parameter \code{b}.
}
\examples{
summary_gamma(a=2, b=4, type="pandoc", style="rmarkdown")
}

