% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/distributions.R
\name{summary_nbinom}
\alias{summary_nbinom}
\title{Summary of a Negative Binomial distribution}
\usage{
summary_nbinom(size, prob, output = "list", ...)
}
\arguments{
\item{size,prob}{parameters of the negative binomial distribution (as for \code{\link[stats]{NegBinomial}})}

\item{output}{\code{"list"} to return a list, \code{"pandoc"} to print a table}

\item{...}{arguments passed to \code{\link[=pander]{pander.data.frame}}}
}
\description{
Mode, mean, variance, and quartiles for a Negative Binomial distribution
with shape parameter \code{size} and probability parameter \code{prob}.
}
\examples{
summary_nbinom(size=2, prob=0.4, output="pandoc", style="rmarkdown")
}

